Add files via upload
INTRODUCTION

British Airways (BA) is the flag carrier airline of the United Kingdom. It is headquartered in London, England, near its main hub at Heathrow airport.
The airline is the second largest UK-based carrier, based on fleet size and passengers carried, behind EasyJet.

METHODOLOGY

R Studio for getting data from website
Web scraping in R using Selector Gadget and Rvest
Data manipulation with Dplyr
Data visualization with R

ANALYSIS AND FINDINGS
1. sentimental analysis know the feeling of the customers ;
From the pie chart, it shows that positive sentiment
Has a percentage of 67.5% while  negative is 32.5%
2. The most used word were, Airway, uncomfortable, airway, good, and seat.

CUSTOMER  BOOKING PREDICTION

1. Multiple Linear Regression was used.
2. Purchase lead was dependent on all other variables in the data set.
with a P-VALUE OF 0.00002 which explains that there is a significant difference with  Purchase Lead. And wouldn't affect the customer's Future booking.